+++ sites/cacademy/CHANGELOG.md
@@ -12,3 +12,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Install richie for C-Academy
 
+### Added
+
+- 🔧(rolling_deploy) Enable the deliver of the static asset files like css,
+  images, fonts, etc. Nginx will also deliver this files, but during rolling
+  deploy, the nginx could have one version of the files and the app still
+  running a different version.