+++ sites/nau/CHANGELOG.md
@@ -16,6 +16,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - ✨(googletagmanager) add Google Tag Manager to nau site without using
   Richie upstream Web Analytics.
+- 🔧(rolling_deploy) Enable the deliver of the static asset files like css,
+  images, fonts, etc. Nginx will also deliver this files, but during rolling
+  deploy, the nginx could have one version of the files and the app still
+  running a different version.
 
 ## [1.13.0] - 2022-06-28